copyWiki

Copies wikis and updates internal links

copyWiki(entity, destinationId, entitySubPageId=NA, destinationSubPageId=NA, updateLinks=TRUE, updateSynIds=TRUE, entityMap=NA)

Arguments

entity

A synapse ID of an entity whose wiki you want to copy

destinationId

Synapse ID of a folder/project that the wiki wants to be copied to

entitySubPageId

Can specify subPageId and copy all of its subwikis
Defaults to None, which copies the entire wiki
subPageId can be found: https://www.synapse.org/#!Synapse:syn123/wiki/1234
In this case, 1234 is the subPageId.

destinationSubPageId

Can specify destination subPageId to copy wikis to
Defaults to None

updateLinks

Update all the internal links. (e.g. syn1234/wiki/34345 becomes syn3345/wiki/49508)
Defaults to True

updateSynIds

Update all the synapse ID's referenced in the wikis. (e.g. syn1234 becomes syn2345)
Defaults to True but needs an entityMap

entityMap

An entity map 'oldSynId','newSynId' to update the synapse IDs referenced in the wiki
Defaults to NA

Details

Value

A list of Objects with three fields: id, title and parentId.

References

Note

See also

Examples

  if (FALSE) {
    copyWiki("syn123", "syn456")
  }