syncToSynapse

Synchronizes files specified in the manifest file to Synapse

syncToSynapse(manifestFile, dryRun=FALSE, sendMessages=TRUE, retries=4)

Arguments

manifestFile

A tsv file with file locations and metadata
to be pushed to Synapse. See below for details

dryRun

Performs validation without uploading if set to True (default is False)

sendMessages

retries

Details

Given a file describing all of the uploads, uploads the content to Synapse and optionally notifies you via Synapse messagging (email) at specific intervals, on errors and on completion.

Manifest File Format:

The format of the manifest file is a tab delimited file with one row per file to upload and columns describing the file. The minimum required columns are path and parent where path is the local file path and parent is the synapse Id of the project or folder where the file is uploaded to. In addition to these columns you can specify any of the parameters to the File constructor (name, synapseStore, contentType) as well as parameters to the syn.store command (used, executed, activityName, activityDescription, forceVersion). Used and executed can be semi-colon (”;”) separated lists of Synapse ids, urls and/or local filepaths of files already stored in Synapse (or being stored in Synapse by the manifest). Any additional columns will be added as annotations.

Required fields:

Field	Meaning	Example

path	local file path or URL	/path/to/local/file.txt

parent	synapse id	syn1235

Common fields:

Field	Meaning	Example

name	name of file in Synapse	Example_file

forceVersion	whether to update version	False

Provenance fields:

Field	Meaning	Example

used	List of items used to generate file	syn1235; /path/to_local/file.txt

executed	List of items exectued	https://github.org/; /path/to_local/code.py

activityName	Name of activity in provenance	“Ran normalization”

activityDescription	Text description on what was done	“Ran algorithm xyx with parameters...”

Annotations:

Any columns that are not in the reserved names described above will be intepreted as annotations of the file.

Other optional fields:

Field	Meaning	Example

synapseStore	Boolean describing whether to upload files	True

contentType	content type of file to overload defaults	text/html

Example manifest file:

path	parent	annot1	annot2	used	executed

/path/file1.txt	syn1243	“bar”	3.1415	“syn124; /path/file2.txt”	“https://github.org/foo/bar“

/path/file2.txt	syn12433	“baz”	2.71	“”	“https://github.org/foo/baz“

Value

References

Note

See also

Examples

  if (FALSE) {
    syncToSynapse("/path/to/manifest.tsv")
  }